show fields of complex type pass nodes
